added new servlet job feature: attach
`attach` makes it possible to attach a job (ZeroVM session) to one of the devices specified in json job config.
It means that the executable and all other objects will be copied to the node that contains that object, and the code will be executed on that node.
Obviously it does make sense only if the device path is a Swift one.

Documentation was also fixed to reflect the new `attach` property.