requirements: add URL for Swift
For some reason, the latest Swift version on PyPI is 1.0.2. However,
the developer do publish pip-installable tarballs on Launchpad:

  https://launchpad.net/swift/

We can instruct Pip to look there with the --find-links flag, and this
flag can be embedded in a requirements.txt file.

This will solve https://github.com/zerovm/zvm-jenkins/issues/22.