zerocloud: avoid swift dependency for gettext (fixes #50)
Importing Gettext from Swift won't work out of the box. Swift relies
on a $SWIFT_LOCALEDIR environment variable being set in order for it
to find its compiled message catalogs (the .mo files). That variable
generally not set when running ZeroCloud.

Furthermore, we have begun marking our own strings for translation
with the _ function. This means that we need our own message catalog
if there is any hope of them being translated.