objectquery: fix content length mismatch cases
Previously, we only checked if the content-length header was not equal
to the amount of content received in a request. This resulted in a "499
Client Closed Request" (indicating a premature disconnect).

This response could be triggered by content either less then OR greater
than that specified in the conteint-length header. This is a bit weird.

This change alters the check and breaks it into two cases:

- actual content < content-length: 499
- actual content > content-length: 400 (Bad Request)

Tests are updated accordingly.