vagrant: improve the local dev env
The git clone on the host is now mapped in to the VM, and changes to the
code in the host dev environment will be reflected inside the VM (after
restarting devstack, of course).

This makes it a lot easier to experiment, debug, and test in a "real"
environment (and not just with the test suite).