vagrant: don't install zerovm-cli or zpm
These packages install oslo.config 1.2.1. The latest devstack needs
oslo.config 1.4, and these two are incompatible.

Technically, we don't really need the clients in the vagrant box
(although if someone wants to run zvsh, etc., inside the VM in addition
to devstack, this will be a problem). To work around this, we just
install python-swiftclient==2.2.0, which installs oslo.config 1.4.x.