proxyquery: fix spurious auth errors on cont. list
An authorization error (presented to the client as a 400 bad request) is
raise in some cases when container listing is done in a chain. A
concrete example:

An api/1.0 application spawns a chained job which does a mapreduce job,
operating a container glob. The glob causes a container listing to be
done, and this is where the authorization fails (if the user is, for
example, anonymous and executing under setuid permission).